Capture and store mouse position in App
Newer versions than the one in this (22.10) version of ubuntu (i.e. the
one currently in homebrew) have mouseX and mouseY fields in the
SDL_MouseWheelEvent struct. What we do instead is store the mouse
position every time we get a mouse motion event and make that position
available as a method in the App class.