Added is_complete() virtual method
Nodes return true here if they are ready to be further processed, and false
if they are not. This is useful in for example a REPL: if the result of a
parse returns false for is_complete(), there is no point in executing the
statement, so for example pressing enter should not try to do that.